more front end changes

+++ ss-tracker/server/models/locations.js
@@ -2,8 +2,8 @@ import mongoose from "mongoose";
 
 const locationsSchema = mongoose.Schema({
     sharkName: String,
-    latitude: Float,
-    longitude: Float,
+    latitude: Float64Array,
+    longitude: Float64Array,
     time: Date
 });
